Slight refactor of sqrl to sql
